server/src/requirements.txt
websockets==13.0.1
yara-python==4.5.1
